requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2024.2.2
coverage==7.4.1
distro==1.9.0
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
inflection==0.5.1
openai==1.12.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
uritemplate==4.1.1
